move helper functions to test_helper.go
